+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Using only the values type, color, firstColor, orientation and secondColor of the "background" to generate the hash keys used for comparison
 
 ## [0.6.1] - 2024-05-27
 - Support theme variables without "$", "$" will be appended when generating the css